+++ pg_search/src/postgres/customscan/mod.rs
@@ -32,6 +32,7 @@ pub mod dsm;
 pub mod exec;
 pub mod explain;
 mod explainer;
+pub mod expr_eval;
 mod hook;
 pub mod joinscan;
 pub mod limit_offset;
@@ -39,6 +40,7 @@ pub mod opexpr;
 pub mod orderby;
 pub mod parallel;
 mod path;
+pub(crate) mod pg_expr_udf;
 pub mod projections;
 pub mod pullup;
 mod pushdown;
